Remove unused files and update dependencies
This commit removes several unused files from the repository, including images, Swift files, configuration files, and workspace files. It also updates the dependencies in the metro.config.js file and the Podfile to ensure compatibility with the latest versions. These changes help clean up the repository and improve the overall performance and stability of the project.